+++ addon/computeds/-base.js
@@ -1,20 +1,9 @@
-import Ember from 'ember';
-
-import getValue from '../utils/get-value';
-import getDependentKeys from '../utils/get-dependent-keys';
-
-const { computed } = Ember;
+import computedUnsafe from 'ember-macro-helpers/computed-unsafe';
 
 export default function computedFactory(fn) {
   return function(...args) {
-    const computedArgs = [].concat(getDependentKeys(args));
-
-    computedArgs.push(function() {
-      const params = args.map((arg) => getValue.call(this, arg));
-
-      return fn.call(this, params);
+    return computedUnsafe(...args, function(...vals) {
+      return fn.call(this, vals);
     });
-
-    return computed(...computedArgs);
   };
 }