+++ addon/computeds/format.js
@@ -1,13 +1,13 @@
 import Ember from 'ember';
 import moment from 'moment';
-import getOwner from 'ember-getowner-polyfill';
+import getOwner from '../utils/get-owner';
 
 import computedFactory from './-base';
 
 const CONFIG_KEY = 'config:environment';
 const { get } = Ember;
 
-export default computedFactory(function formatComtputed([value, optionalFormat]) {
+export default computedFactory(function formatComputed([value, optionalFormat]) {
   if (!optionalFormat) {
     const owner = getOwner(this);