+++ addon/helpers/is-between.js
@@ -1,15 +1,18 @@
 import Ember from 'ember';
-import moment from 'moment';
 
 import computeFn from '../utils/helper-compute';
 import BaseHelper from './-base';
 
 export default BaseHelper.extend({
+  moment: Ember.inject.service(),
+
   globalAllowEmpty: false,
 
   compute: computeFn(function(params, { precision, inclusivity, locale, timeZone }) {
     this._super(...arguments);
 
+    const moment = this.get('moment');
+    const _params = [].concat(params);
     const { length } = params;
 
     if (length < 2 || length > 3) {
@@ -17,14 +20,11 @@ export default BaseHelper.extend({
     }
 
     const args = [];
-    const comparisonArgs = Ember.A();
 
     if (length > 2) {
-      args.push(params.shift());
+      args.push(_params.shift());
     }
 
-    comparisonArgs.pushObjects(params);
-
-    return this.morphMoment(moment(...args), { locale, timeZone }).isBetween(...comparisonArgs, precision, inclusivity);
+    return this.morphMoment(moment.moment(...args), { locale, timeZone }).isBetween(..._params, precision, inclusivity);
   })
 });