+++ addon/services/moment.js
@@ -1,9 +1,9 @@
 import Ember from 'ember';
 import moment from 'moment';
 
-const { computed, Logger:logger } = Ember;
+const { computed, get, getProperties, set, Logger:logger } = Ember;
 
-export default Ember.Service.extend({
+export default Ember.Service.extend(Ember.Evented, {
   _timeZone: null,
 
   locale: null,
@@ -11,29 +11,40 @@ export default Ember.Service.extend({
 
   timeZone: computed('_timeZone', {
     get() {
-      return this.get('_timeZone');
+      return get(this, '_timeZone');
     },
 
     set(propertyKey, timeZone) {
       if (!moment.tz) {
-        logger.warn('[ember-moment] attempted to set timezone, but moment-timezone unavailable.');
+        logger.warn('[ember-moment] attempted to set timezone, but moment-timezone is not setup.');
         return;
       }
 
-      this.set('_timeZone', timeZone);
+      set(this, '_timeZone', timeZone);
 
       return timeZone;
     }
   }),
 
+  setLocale(locale) {
+    this.changeLocale(locale);
+  },
+
   changeLocale(locale) {
-    this.set('locale', locale);
+    set(this, 'locale', locale);
+    moment.locale(locale);
+    this.trigger('localeChanged', locale);
+  },
+
+  setTimeZone(timeZone) {
+    this.changeTimeZone(timeZone);
   },
 
   changeTimeZone(timeZone) {
-    this.set('timeZone', timeZone);
+    set(this, 'timeZone', timeZone);
+    this.trigger('timeZoneChanged', timeZone);
   },
-  
+
   isMoment(obj) {
     return moment.isMoment(obj);
   },
@@ -64,7 +75,7 @@ export default Ember.Service.extend({
 
     return time;
   },
-  
+
   getInstance() {
     const locale = this.get('locale');
     if (locale) {