+++ app/helpers/moment-from-now.js
@@ -1,7 +1,7 @@
 import Ember from 'ember';
 import config from '../config/environment';
-import DurationHelper from 'ember-moment/helpers/moment-from-now';
+import FromNowHelper from 'ember-moment/helpers/moment-from-now';
 
-export default DurationHelper.extend({
+export default FromNowHelper.extend({
   globalAllowEmpty: !!Ember.get(config, 'moment.allowEmpty')
 });