+++ CHANGELOG.md
@@ -1,12 +1,49 @@
-### 7.0.0-beta.2
-* now helper returns `moment.now()` instead of `new Date()`
+### 7.3.0
 
-### 7.0.0-beta.1
+* Setting locale now sets locale on global moment object
+* Added `setLocale` and `setTimeZone`.  Better naming.  Will continue to support `changeLocale` and `changeTimeZone` but have updated README to prefer new method names
+* Added `localeChanged` and `timeZoneChanged` events
+* @mfeltz scoped moment-subtract and moment-add to use the moment service `locale` property
 
-* Upgrade ember-cli-moment-shim to 1.2.0
-* Add moment query helpers (is-between, is-sae-or-after, is-same-or-before, is-same, is-after, is-before)
+### 7.2.0
+
+* @kellyselden add back `ember-macro-helpers`
+
+### 7.1.1
+
+* Removed `ember-macro-helpers`
+
+### 7.1.0
+
+* @kellyselden ported computed macro factory to use computed macro utility methods from `ember-macro-helpers`
+
+### 7.0.3
+
+* Upgrade `ember-cli-moment-shim` to 3.0.0
+* now helper now recomputes using setTimeout instead of run.later #205
+
+### 7.0.2
+
+* Revert upgrade `ember-cli-moment-shim` to 2.2.1 (moment.now issue)
+
+### 7.0.1
+
+* Upgrade `ember-cli-moment-shim` to 2.2.1
+
+### 7.0.0
+
+* Upgrade ember-cli-moment-shim to 1.3.0
+* `now` helper returns `moment.now()` instead of `new Date()`
+* Added `moment` helper
+* Added `moment-calendar` helper
+* @GarPit Added `moment-to-date` helper
+* @stavarotti added `unix` helper
+* @yads added `moment-subtract` `moment-add` helpers
+* Removed `ember-getowner-polyfill` as a direct dep
+* Add moment query helpers (`is-between`, `is-same-or-after`, `is-same-or-before`, `is-same`, `is-after`, `is-before`)
 * BREAKING: no longer defaults moment-format to LLLL and instead uses the moment default format (ISO 8601)
   * Applications will need to specify the default format to LLLL if they want to keep this behavior: https://github.com/stefanpenner/ember-moment#global-default-output-format
+* Adopted yarn
 
 ### 6.0.0